requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
MarkupSafe==3.0.2
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
Werkzeug==3.1.3
reportlab
python-barcode